+++ libs/components/core/testing/src/shared/harness-filters.ts
@@ -1,8 +1,11 @@
 import { BaseHarnessFilters } from '@angular/cdk/testing';
 
 /**
- * @experimental
+ * A set of criteria that can be used to filter a list of `SkyComponentHarness` instances.
  */
 export interface SkyHarnessFilters extends BaseHarnessFilters {
+  /**
+   * Only find instances whose `data-sky-id` attribute matches the given value.
+   */
   dataSkyId?: string | RegExp;
 }