+++ libs/components/forms/testing/src/input-box/input-box-harness.ts
@@ -1,0 +1,34 @@
+import {
+  ComponentHarness,
+  HarnessPredicate,
+  HarnessQuery,
+} from '@angular/cdk/testing';
+import { SkyComponentHarness } from '@skyux/core/testing';
+
+import { SkyInputBoxHarnessFilters } from './input-box-harness-filters';
+
+/**
+ * Harness for interacting with an input box component in tests.
+ */
+export class SkyInputBoxHarness extends SkyComponentHarness {
+  public static hostSelector = 'sky-input-box';
+
+  /**
+   * Gets a `HarnessPredicate` that can be used to search for a
+   * `SkyInputBoxHarness` that meets certain criteria.
+   */
+  public static with(
+    filters: SkyInputBoxHarnessFilters
+  ): HarnessPredicate<SkyInputBoxHarness> {
+    return SkyInputBoxHarness.getDataSkyIdPredicate(filters);
+  }
+
+  /**
+   * Returns a child harness.
+   */
+  public async queryHarness<T extends ComponentHarness>(
+    query: HarnessQuery<T>
+  ): Promise<T | null> {
+    return this.locatorForOptional(query)();
+  }
+}