+++ libs/components/lookup/src/lib/modules/autocomplete/types/autocomplete-search-function.ts
@@ -1,6 +1,8 @@
+import { SkyAutocompleteSearchArgs } from './autocomplete-search-args';
 import { SkyAutocompleteSearchFunctionResponse } from './autocomplete-search-function-response';
 
 export type SkyAutocompleteSearchFunction = (
   searchText: string,
-  data: any[]
+  data: any[],
+  args?: SkyAutocompleteSearchArgs
 ) => SkyAutocompleteSearchFunctionResponse;