+++ libs/components/lookup/testing/src/public-api.ts
@@ -1,5 +1,12 @@
+export * from './autocomplete/autocomplete-harness';
+export * from './autocomplete/autocomplete-harness-filters';
+export * from './autocomplete/autocomplete-search-result-harness';
+export * from './autocomplete/autocomplete-search-result-harness-filters';
+
 export * from './country-field/country-field-fixture';
 export * from './country-field/country-field-testing.module';
 
 export * from './search/search-fixture';
+export * from './search/search-harness';
+export * from './search/search-harness-filters';
 export * from './search/search-testing.module';